refine file for merge
